BlowOff &&Gauge

+++ AnaphaseTeam/AnaphaseTeam/source/scene/each/GamePlay.cpp
@@ -54,7 +54,7 @@ void GamePlay::update(float deltaTime)
 	m_enemys.collision(m_player);
 	m_cameracontroller.update(deltaTime);
 	
-	for (int i = 0; i < 2- m_enemys.size(); i++)
+	for (int i = 0; i < 2- static_cast<int>(m_enemys.size()); i++)
 	{
 		Math::Random rnd;
 		Enemy* e = new Enemy(Transform(0, { 0,0,0 }, { rnd(-10.0f,10.0f),0,rnd(-10.0f,10.0f) }));
